titan_social/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
pytesseract>=0.3.10
replicate>=0.25.0

# UI
streamlit>=1.36.0
